+++ __test__/components/quote_stats.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as renderer from 'react-test-renderer';
 
-import { QuoteStatsComponent } from '@components/QuoteStatsComponent';
+import { QuoteStatsComponent } from '@components/quote_stats';
 
 describe('Component: QuoteStats', (): void => {
   it('will be rendered with no error', (): void => {
